major refactoring done
