+++ src/nn/activation/map.py
@@ -1,0 +1,9 @@
+from nn.activation.functions import Sigmoid, Softmax, ReLu, Identity, Tanh
+
+activation_func_map = {
+    "sigmoid": Sigmoid(),
+    "softmax": Softmax(),
+    "ReLU": ReLu(),
+    "identity": Identity(),
+    "tanh": Tanh()
+}
