+++ src/nn/loss/map.py
@@ -1,0 +1,5 @@
+from nn.loss.functions import CrossEntropy
+
+loss_func_map = {
+        "cross_entropy": CrossEntropy()
+    }
